Change package, couchebochebase to sql
