+++ mobile/src/main/java/cz/uhk/fim/kikm/wearnavigation/MainActivity.java
@@ -1,0 +1,26 @@
+package cz.uhk.fim.kikm.wearnavigation;
+
+import android.os.Bundle;
+
+public class MainActivity extends BaseActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    protected void updateUI() {
+        // Not used
+    }
+
+    @Override
+    protected int getContentViewId() {
+        return R.layout.activity_main;
+    }
+
+    @Override
+    protected int getNavigationMenuItemId() {
+        return R.id.action_show_main;
+    }
+}
