+++ mobile/src/main/java/cz/uhk/fim/kikm/wearnavigation/activities/databaseTest/CouchebaseActivity.java
@@ -1,0 +1,15 @@
+package cz.uhk.fim.kikm.wearnavigation.activities.databaseTest;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import cz.uhk.fim.kikm.wearnavigation.R;
+
+public class CouchebaseActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_couchebase);
+    }
+}
