--- mobile/src/main/java/cz/uhk/fim/kikm/wearnavigationsimple/MainActivity.java
@@ -1,29 +1,0 @@
-package cz.uhk.fim.kikm.wearnavigationsimple;
-
-import android.os.Bundle;
-
-import cz.uhk.fim.kikm.wearnavigationsimple.activities.bluetoothTest.BluetoothTest;
-import cz.uhk.fim.kikm.wearnavigationsimple.activities.devices.ShowDevicesActivity;
-
-public class MainActivity extends BaseActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    protected void updateUI() {
-        // Not used
-    }
-
-    @Override
-    protected int getContentViewId() {
-        return R.layout.activity_main;
-    }
-
-    @Override
-    protected int getNavigationMenuItemId() {
-        return R.id.action_show_main;
-    }
-}
