+++ mobile/src/main/res/layout/activity_bluetooth_test.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cz.uhk.fim.kikm.wearnavigationsimple.activities.bluetoothTest.BluetoothTest">
+    tools:context="cz.uhk.fim.kikm.wearnavigation.activities.bluetoothTest.BluetoothTest">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/abt_message_list"
